Chat-with-DB-Deployment/pyproject.toml
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-with-db"
version = "0.1.0"
description = "A natural language to SQL query system with visualization capabilities"
authors = [
    {name = "Arjun Seeramsetty", email = "arjun.seeramsetty@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "structlog==23.2.0",
    "openai==1.3.7",
    "numpy==1.24.3",
    "pandas==2.0.3",
    "scikit-learn==1.3.0",
    "faiss-cpu==1.7.4",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "httpx==0.25.2",
    "aiofiles==23.2.1",
    "sqlglot==19.0.0",
    "sqlfluff==2.3.0",
    "sqlparse==0.4.4",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "mypy",
    "types-PyYAML==6.0.12.20250516",
    "types-python-dateutil==2.9.0.20250708",
    "types-requests==2.32.4.20250611",
    "PyYAML==6.0.1",
    "python-dateutil==2.8.2",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["tests*", "agent-ui*", "config*"] 